# web/openapi/raffa-api.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Raffa API",
    "version": "1.1.0",
...
  },
  "servers": [
    {
      "url": "/",
      "description": "Relative to the runtime-configured API origin (AppConfig.apiBaseUrl, ADR-012 \"config, not code\"; see web/src/config/appConfig.ts). This document intentionally does not hard-code a dev/demo host."
    }
  ],
  "paths": {
...
    "/api/chat/query": {
      "post": {
        "operationId": "askRaffa",
        "summary": "Ask Raffa: route a natural-language question and answer it with citations, or abstain.",
        "description": "Maps to PostChatQueryAsync in backend/src/Raffa.Api/ChatEndpointExtensions.cs (task E02/F04/US01/T01 query-router, E02/F04/US02/T01 rag-citations, E02/F04/US02/T02 abstain-guard; product spec §8.3/§8.4). Task E07/F04/US01/T01 (ask-raffa-ui, ADR-020 screen 7) adds this operation -- see this document's own info.description for the full provenance/gap notes (Structured-not-wired message, Semantic-abstain reason, and the Document-vs-Clause citation-lookup gap). Same interim X-Tenant-Id header placeholder as every other operation above.",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "question"
                ],
                "properties": {
                  "question": {
                    "type": "string",
                    "description": "The user's natural-language question. A blank/whitespace-only value is rejected as 400."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Always 200 once routing succeeds -- the Structured-not-wired and the Semantic answer/abstain outcomes are both valid, expected shapes of the same envelope, never an error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "question",
                    "intent",
                    "canDetermine",
                    "answer",
                    "citations",
                    "message"
                  ],
                  "properties": {
                    "question": {
                      "type": "string"
                    },
                    "intent": {
                      "type": "string",
                      "enum": [
                        "Structured",
                        "Semantic"
                      ]
                    },
                    "canDetermine": {
                      "type": "boolean"
                    },
                    "answer": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "citations": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "documentId",
                          "page",
                          "section"
                        ],
                        "properties": {
                          "documentId": {
                            "type": "string",
                            "description": "Composite SourceType:SourceId (e.g. Document:<guid> or Clause:<guid>) -- see this document's own info.description."
                          },
                          "page": {
                            "type": [
                              "integer",
                              "null"
                            ],
                            "description": "Always null today -- Embedding carries no page column yet (ChatEndpointExtensions.ToEvidenceSnippet)."
                          },
                          "section": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "description": "Real chunk position (e.g. 'chunk 2'), not a document section number."
                          }
                        }
                      }
                    },
                    "message": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "description": "Non-null only for the Structured-not-wired case -- see this document's own info.description."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing/invalid X-Tenant-Id header, a blank question, or a retrieval/answer failure (EmbeddingRetrievalService/RagAnswerService Result.Error).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/conversations/{id}/messages": {
      "post": {
        "operationId": "postConversationMessage",
        "summary": "Asks a question in an existing conversation; runs the Ask engine and returns the reply.",
        "description": "Maps to PostConversationMessageAsync in backend/src/Raffa.Api/ConversationsEndpointExtensions.cs -- the one endpoint that actually runs the Ask engine (Raffa.Api.AskCopilotService.AskAsync, task E13/F06/US01/T01), then persists both the caller's question and Raffa's reply (ConversationService.AppendMessageAsync) before wire-shaping the same object back (shared with POST /api/chat/query's own alias, ConversationsEndpointExtensions.AskAndAppendAsync). Request body `{ question }` (ConversationsEndpointExtensions.PostConversationMessageRequest, hand-written client-side, see this document's 'Codegen tool choice' paragraph). Reply shape is ADR-024 §6 / inputs/requirements.md §6, kind decides the layout: `answer` (markdown + citations + actions + followUps), `abstain` (a true-insufficiency reply, citations/actions/followUps empty), `redirect` (greeting/off-domain/needs_document, one CTA), `refusal` (legal, one CTA). `actions[].kind` is 'navigate' | 'upload' (Raffa.Chat.Application.Capabilities.CopilotActionKind) -- inputs/requirements.md §6's own JSON example shows 'primary'/'secondary' instead, an illustrative visual-priority label the real backend never emits; web/src/routes/ask/askViewModel.ts maps the real navigate/upload/index-ordered actions onto ReplyAction's presentational primary/secondary at the UI boundary (replyTypes.ts, task E13/F09/US01/T02, is out of this task's file scope to change). See GET /api/conversations/{id} above for why `corpus` is a plain string, not a 3-value enum. 404 under the identical tenant/user/unknown-id rule GET /api/conversations/{id} documents.",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "See GET /api/conversations above."
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "Conversation.Id."
          }
        ],
        "responses": {
          "200": {
            "description": "The routed reply -- always 200 once the conversation is confirmed to exist, per ChatEndpointExtensions' own long-standing convention for this same composition root; an abstain/redirect/refusal is a successful, honest answer, not an error status.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "conversationId", "messageId", "kind", "answerMarkdown",
                    "citations", "actions", "provenance", "followUps"
                  ],
                  "properties": {
                    "conversationId": { "type": "string", "format": "uuid" },
                    "messageId": { "type": "string", "format": "uuid" },
                    "kind": { "type": "string", "enum": ["answer", "abstain", "redirect", "refusal"] },
                    "answerMarkdown": {
                      "type": "string",
                      "description": "Narrated prose with inline [n] markers; empty for abstain/redirect/refusal? no -- redirect/refusal still carry warm prose here (R-ASK-07); only truly nothing to narrate is empty."
                    },
                    "citations": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "n", "corpus", "title", "subtitle", "snippet", "documentId",
                          "contractId", "page", "section", "previewUrl", "href", "recordId"
                        ],
                        "properties": {
                          "n": { "type": "integer" },
                          "corpus": {
                            "type": "string",
                            "description": "See GET /api/conversations/{id} above's identical citations[].corpus for the full 'tenant/market/raffa, occasionally calc' provenance note."
                          },
                          "title": { "type": "string" },
                          "subtitle": { "type": ["string", "null"] },
                          "snippet": { "type": "string" },
                          "documentId": { "type": ["string", "null"] },
                          "contractId": { "type": ["string", "null"], "format": "uuid" },
                          "page": { "type": ["integer", "null"] },
                          "section": { "type": ["string", "null"] },
                          "previewUrl": { "type": ["string", "null"] },
                          "href": {
                            "type": ["string", "null"],
                            "description": "A tenant citation's own destination -- observed backend behaviour (AskCopilotService.cs's PackItem constructions) always sets this to the bare `/contracts/{contractId}`, never `/contracts/{contractId}?clause=...`, despite inputs/requirements.md §6's own illustrative example showing a `?clause=` suffix. web/src/routes/ask/askViewModel.ts appends `?page={page}` itself (when `page` is present) before navigating, reusing web/src/routes/contracts/contract360/contract360ViewModel.ts#resolveHighlightedClauseId's existing page-based clause lookup -- a `?clause=` deep link cannot be built client-side today because no citation field carries a raw clauseId (R-ASK-08 'no guids ever rendered' -- consistent with that, not a client oversight)."
                          },
                          "recordId": { "type": ["string", "null"] }
                        }
                      }
                    },
                    "actions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["label", "href", "kind"],
                        "properties": {
                          "label": { "type": "string" },
                          "href": { "type": "string" },
                          "kind": { "type": "string", "enum": ["navigate", "upload"] }
                        }
                      }
                    },
                    "provenance": {
                      "type": "object",
                      "required": ["sources", "modelId", "promptVersion", "inputHash"],
                      "properties": {
                        "sources": { "type": "array", "items": { "type": "string" } },
                        "modelId": { "type": ["string", "null"] },
                        "promptVersion": { "type": ["string", "null"] },
                        "inputHash": { "type": ["string", "null"] }
                      }
                    },
                    "followUps": { "type": "array", "items": { "type": "string" } }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing/invalid X-Tenant-Id or X-User-Id header, a non-GUID id in the route, or a blank 'question'.",
            "content": { "application/json": { "schema": { "type": "string" } } }
          },
          "404": {
            "description": "No such conversation for this tenant and caller."
          }
        }
      }
    },
...
  }
}